Time: 40 ms (75.85%), Space: 125 MB (81.16%) - LeetHub

+++ 2300-successful-pairs-of-spells-and-potions/2300-successful-pairs-of-spells-and-potions.cpp
@@ -5,7 +5,7 @@ public:
         sort(potions.begin(),potions.end());
         vector<int> ans(n,0);
         for(int i=0;i<n;i++){
-            int minPotion=((success+spells[i]-1)/spells[i]);
+            long long minPotion=((success+spells[i]-1)/spells[i]);
             auto it=lower_bound(potions.begin(),potions.end(),minPotion);
             if(it==potions.end()){
                 continue;
